Handle "raw" messages
This is basically just an extra field on messages that we thread through and
then don't display. The intent is to use it for systems that implement their own
native serialization and don't want to rely on JSON. These systems can use the
JSON display for any debugging information they want, but don't have to read it
back in.